renamed pkg/insert to pkg/init.
