+++ pkg/init/genbank.go
@@ -1,0 +1,1 @@
+package init
