+++ pkg/init/rhea_test.go
@@ -1,4 +1,4 @@
-package insert
+package init
 
 import (
 	"context"
